[FE] 폴더정리 반영
