+++ frontend/package.json
@@ -16,8 +16,9 @@
     "jetifier": "1.6.6",
     "react": "17.0.1",
     "react-native": "0.64.0",
+    "react-native-fade": "2.0.2",
     "react-native-gesture-handler": "1.10.3",
-    "react-native-reanimated": "2.0.1",
+    "react-native-reanimated": "2.1.0",
     "react-native-safe-area-context": "3.2.0",
     "react-native-screens": "2.18.1",
     "rn-bottom-drawer": "1.4.3",
@@ -28,6 +29,7 @@
     "@babel/runtime": "7.13.10",
     "@react-native-community/eslint-config": "2.0.0",
     "babel-jest": "26.6.3",
+    "babel-plugin-module-resolver": "4.1.0",
     "eslint": "7.14.0",
     "jest": "26.6.3",
     "metro-react-native-babel-preset": "0.64.0",
