+++ frontend/src/components/WhiteInput.js
