+++ frontend/src/pages/Login/EmailJoinScreen.js
@@ -1,15 +1,12 @@
 import React, { Component } from 'react';
 import { StyleSheet, Text, View, Image, TextInput } from 'react-native';
-import BasicButton from '../../components/BasicButton';
 import SimpleInput from '../../components/SimpleInput'
-import BasicTitle from '../../components/BasicTitle'
 
 function EmailJoinScreen({ navigation }) {
   return (
     <View style={styles.container}>
           <View style={styles.content}>
               <SimpleInput onSubmitEditing ={() => navigation.navigate('SendEmail')}>
-                  
               </SimpleInput>
           </View>
     </View>
