+++ frontend/src/pages/Login/SendEmailScreen.js
@@ -1,26 +1,20 @@
 import React, { Component } from 'react';
 import { StyleSheet, Text, View, Image } from 'react-native';
-import BasicButton from '../../components/BasicButton';
-import LoginDrawer from '../../components/LoginDrawer'
 
-
-function StartScreen({ navigation }) {
+function SendEmailScreen({ navigation }) {
   return (
     <View style={styles.container}>
       <View style={styles.header} />
-      <View style={styles.title}>
-        <Text style={{ fontSize: 35, color: 'white' }}>후훗{'\n'}블루파-렌탈을{'\n'}시작한다</Text>
-      </View>
+      
       <View style={styles.content}>
         <Image
-          style={{ height: '100%', width: '100%', resizeMode: 'contain' }}
-          source={require('../../assets/start_img.jpg')} />
+          style={{ tintColor: 'white',height: '100%', width: '100%', resizeMode: 'contain' }}
+          source={require('../../assets/mail.png')} />
+      </View>
+      <View style={styles.title}>
+        <Text style={{ fontSize: 35, color: 'white' }}>이메일을 확인해주세요!</Text>
       </View>
       <View style={styles.footer}>
-        <BasicButton
-          buttonColor={'#444'}
-          title={'시작하기'}
-          onPress={() => navigation.navigate('Join')} />
       </View>
 
     </View>
@@ -56,4 +50,4 @@ const styles = StyleSheet.create({
     width: 150,
   }
 });
-export default StartScreen;
+export default SendEmailScreen;
